+++ dbFuntions.py
@@ -1,0 +1,20 @@
+import csv
+import os
+# Función para escribir en un archivo CSV
+def escribir_csv(filename, data):
+    fieldnames = data[0].keys() if data else []
+    with open(filename, mode='w', newline='') as file:
+        writer = csv.DictWriter(file, fieldnames=fieldnames)
+        if fieldnames:
+            writer.writeheader()
+            writer.writerows(data)
+
+
+# Función para leer un archivo CSV
+def leer_csv(filename):
+    if not os.path.exists(filename):
+        return []
+    with open(filename, mode='r', newline='') as file:
+        reader = csv.DictReader(file)
+        return list(reader)
+
